pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "neu-users"
dynamic = ["version"]
description = "Users microservice for Neu"
readme = "README.md"
requires-python = ">=3.10"
license = "GPL-3.0-or-later"
authors = [
  { name = "Viktor Karamanis", email = "viktor.karamanis@outlook.com" },
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Operating System :: OS Independent",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "fastapi[standard]>=0.115.4,<1.0.0",
  "python-dotenv>=1.0.1,<2.0.0",
  "pydantic>2.0.0,<3.0.0",
  "email-validator>=2.2.0,<3.0.0",
  "redis>=5.2.0,<6.0.0",
  "redis-om>=0.3.3,<1.0.0",
  "isodate>=0.7.2,<1.0.0",
  "neu-sdk @ git+https://github.com/poulet-lab/neu-sdk.git@0.1.4",
]

[project.urls]
Homepage = "https://github.com/poulet-lab/neu-users"
Documentation = "https://github.com/poulet-lab/neu-users"
Changelog = "https://github.com/poulet-lab/neu-users/release-notes/"
Source = "https://github.com/poulet-lab/neu-users"


[tool.hatch.version]
path = "neu_users/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["neu_users"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.envs.lint]
detached = true
dependencies = ["mypy", "ruff"]

[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:neu_users tests}"
style = "ruff check"
fmt = "ruff format"
all = ["style", "typing"]

[tool.mypy]
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 120
indent-width = 4
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]

[tool.ruff.lint]
select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
ignore = ["ISC001"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402", "PLR2004", "S101", "TID252"]

[tool.ruff.format]
docstring-code-format = true
line-ending = "auto"
docstring-code-line-length = "dynamic"

[tool.ruff.lint.isort]
known-first-party = ["neu_users"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[[tool.hatch.envs.hatch-test.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.dev.scripts]
run = "uvicorn neu_users.main:app --reload"

[tool.coverage.run]
source_pkgs = ["neu_users", "tests"]
branch = true
parallel = true

[tool.coverage.paths]
neu_users = ["neu_users", "*/neu-users/neu_users"]
tests = ["tests", "*/neu-users/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]
